refactor: - 프로젝트 일관성을 위한 global 폴더에서 계층 분리 - 분리대상: 커스텀에러, 공통 응답, 전역 예외 처리기
